+++ apps/ledger-live-desktop/src/newArch/Collectibles/components/DetailDrawer/components/PanAndZoom.tsx
@@ -71,6 +71,7 @@ const PanAndZoomBody = memo<PanAndZoomBodyProps>(
           uri={imageUri}
           mediaType={mediaType}
           contentType={contentType}
+          isLoading={false}
         />
       </PrismaZoom>
     </ImageContainer>