+++ apps/ledger-live-desktop/src/newArch/Collectibles/types/Media.ts
@@ -16,21 +16,23 @@ type BaseMediaProps = BaseProps & {
   previewUri?: string | undefined;
   originalUri?: string | undefined;
   mediaType: string | undefined;
-  setUseFallback: (a: boolean) => void;
+  setUseFallback?: (useFallback: boolean) => void;
 };
 
 export type ImageProps = BaseMediaProps & {
   collectibleName?: string | null;
-  onClick?: (e: React.MouseEvent) => void;
+  onClick?: (event: React.MouseEvent) => void;
   isFallback: boolean;
+  isLoading: boolean;
 };
 
 export type VideoProps = BaseMediaProps;
 
 export type MediaProps = BaseMediaProps & {
+  isLoading: boolean;
   useFallback: boolean;
   contentType: string | undefined;
   collectibleName?: string | null;
   mediaFormat?: keyof NFTMedias;
-  onClick?: (e: React.MouseEvent) => void;
+  onClick?: (event: React.MouseEvent) => void;
 };