+++ apps/ledger-live-desktop/src/renderer/screens/exchange/Swap2/Form/SwapWebView.tsx
@@ -18,6 +18,7 @@ import { TopBar } from "~/renderer/components/WebPlatformPlayer/TopBar";
 import useTheme from "~/renderer/hooks/useTheme";
 import {
   counterValueCurrencySelector,
+  discreetModeSelector,
   enablePlatformDevToolsSelector,
   languageSelector,
 } from "~/renderer/reducers/settings";
@@ -100,6 +101,7 @@ const SwapWebView = ({
   const webviewAPIRef = useRef<WebviewAPI>(null);
   const { setDrawer } = React.useContext(context);
   const [webviewState, setWebviewState] = useState<WebviewState>(initialWebviewState);
+  const discreetMode = useSelector(discreetModeSelector);
   const fiatCurrency = useSelector(counterValueCurrencySelector);
   const locale = useSelector(languageSelector);
   const redirectToHistory = useRedirectToSwapHistory();
@@ -273,6 +275,7 @@ const SwapWebView = ({
       provider: swapState?.provider,
       to: targetCurrency?.id,
       toAccountId: swapState?.toAccountId,
+      fromAccountId: swapState?.fromAccountId,
     };
 
     Object.entries(swapParams).forEach(([key, value]) => {
@@ -295,6 +298,7 @@ const SwapWebView = ({
     swapState?.loading,
     swapState?.provider,
     swapState?.toAccountId,
+    swapState?.fromAccountId,
     targetCurrency?.id,
   ]);
 
@@ -338,6 +342,7 @@ const SwapWebView = ({
             theme: themeType,
             lang: locale,
             currencyTicker: fiatCurrency.ticker,
+            discreetMode,
           }}
           onStateChange={onStateChange}
           ref={webviewAPIRef}