+++ apps/ledger-live-mobile/src/const/navigation.ts
@@ -510,6 +510,12 @@ export enum ScreenName {
   WalletSyncActivationSettings = "WalletSyncActivationSettings",
   MockedAddAssetButton = "MockedAddAssetButton",
   GenericLandingPage = "GenericLandingPage",
+
+  // Web3Hub
+  Web3HubMain = "Web3HubMain",
+  Web3HubSearch = "Web3HubSearch",
+  Web3HubTabs = "Web3HubTabs",
+  Web3HubApp = "Web3HubApp",
 }
 
 export enum NavigatorName {
@@ -620,4 +626,7 @@ export enum NavigatorName {
   AnalyticsOptInPrompt = "AnalyticsOptInPrompt",
   WalletSyncActivationSettings = "WalletSyncActivationSettings",
   LandingPages = "LandingPages",
+
+  // Web3Hub
+  Web3Hub = "Web3Hub",
 }