+++ libs/coin-modules/coin-tezos/src/bridge/index.ts
@@ -3,6 +3,7 @@ import {
   makeAccountBridgeReceive,
   makeScanAccounts,
 } from "@ledgerhq/coin-framework/bridge/jsHelpers";
+import { CoinConfig } from "@ledgerhq/coin-framework/config";
 import { SignerContext } from "@ledgerhq/coin-framework/signer";
 import type { CurrencyBridge, AccountBridge } from "@ledgerhq/types-live";
 import getAddressWrapper from "@ledgerhq/coin-framework/bridge/getAddressWrapper";
@@ -17,6 +18,7 @@ import { createTransaction } from "./createTransaction";
 import { buildSignOperation } from "./signOperation";
 import signerGetAddress from "../signer";
 import { broadcast } from "./broadcast";
+import tezosCoinConfig, { TezosCoinConfig } from "../config";
 
 function buildCurrencyBridge(signerContext: SignerContext<TezosSigner>): CurrencyBridge {
   const getAddress = signerGetAddress(signerContext);
@@ -57,7 +59,12 @@ function buildAccountBridge(
   };
 }
 
-export function createBridges(signerContext: SignerContext<TezosSigner>) {
+export function createBridges(
+  signerContext: SignerContext<TezosSigner>,
+  coinConfig: CoinConfig<TezosCoinConfig>,
+) {
+  tezosCoinConfig.setCoinConfig(coinConfig);
+
   return {
     currencyBridge: buildCurrencyBridge(signerContext),
     accountBridge: buildAccountBridge(signerContext),