+++ libs/ledgerjs/packages/errors/src/index.ts
@@ -94,6 +94,7 @@ export const NoAccessToCamera = createCustomErrorClass("NoAccessToCamera");
 export const NotEnoughGas = createCustomErrorClass("NotEnoughGas");
 // Error message specifically for the PTX swap flow
 export const NotEnoughGasSwap = createCustomErrorClass("NotEnoughGasSwap");
+export const TronEmptyAccount = createCustomErrorClass("TronEmptyAccount");
 export const NotSupportedLegacyAddress = createCustomErrorClass("NotSupportedLegacyAddress");
 export const GasLessThanEstimate = createCustomErrorClass("GasLessThanEstimate");
 export const PriorityFeeTooLow = createCustomErrorClass("PriorityFeeTooLow");
@@ -157,8 +158,19 @@ export const SyncError = createCustomErrorClass("SyncError");
 export const PairingFailed = createCustomErrorClass("PairingFailed");
 export const PeerRemovedPairing = createCustomErrorClass("PeerRemovedPairing");
 export const GenuineCheckFailed = createCustomErrorClass("GenuineCheckFailed");
-export const LedgerAPI4xx = createCustomErrorClass("LedgerAPI4xx");
-export const LedgerAPI5xx = createCustomErrorClass("LedgerAPI5xx");
+type NetworkType = {
+  status: number;
+  url: string | undefined;
+  method: string;
+};
+export const LedgerAPI4xx = createCustomErrorClass<
+  NetworkType,
+  LedgerErrorConstructor<NetworkType>
+>("LedgerAPI4xx");
+export const LedgerAPI5xx = createCustomErrorClass<
+  NetworkType,
+  LedgerErrorConstructor<NetworkType>
+>("LedgerAPI5xx");
 export const FirmwareOrAppUpdateRequired = createCustomErrorClass("FirmwareOrAppUpdateRequired");
 
 // SpeedUp / Cancel EVM tx