+++ libs/trustchain/src/__tests__/sdk/randomMemberTryToDestroy.ts
@@ -1,0 +1,4 @@
+import json from "./randomMemberTryToDestroy.json";
+import { replayTrustchainSdkTests } from "../../test-helpers/replayTrustchainSdkTests";
+import { scenario } from "../../test-scenarios/randomMemberTryToDestroy";
+replayTrustchainSdkTests(json, scenario);